Функция генератор рандомной строки. Принимает 2 аргумента, 1 - целочисленный, 2 - строку
