+++ .gitignore
@@ -1,1 +1,2 @@
 static/bundle.js
+.idea/
