+++ server/src/routes/__init__.py
@@ -5,3 +5,8 @@ blueprint = Blueprint('routes', __name__)
 @blueprint.route('/login')
 def login():
   return 'login'
+
+
+@blueprint.route('/config')
+def config():
+  return 'config'
